Add files via upload
This project aims to build an effective credit card fraud detection system using Python. We leveraged various libraries including pandas for data manipulation, numpy for numerical operations, matplotlib and seaborn for data visualization, and scikit-learn (sklearn) for machine learning tasks. To address the significant class imbalance in the dataset, the imblearn module was employed for data balancing, ensuring that the model is robust and unbiased in predicting fraudulent transactions.