Add files via upload
This project focuses on the classification of the Iris dataset, a well-known dataset in machine learning, to predict the species of an iris flower based on its morphological features. The Iris dataset includes measurements for three types of iris flowers: Setosa, Versicolor, and Virginica, with features such as petal length, petal width, sepal length, and sepal width.